Bug Fix - User Input Method Calls
- Bug:
Fixed an issue which was causing the get_user_keyword method to be called upon either cipher selection. There was no logic in that, or in get_user_offset , which checked the user_cipher_type.

- Fix:
Added an if statement to get_user_keyword and get_user_offset which checks the user_cipher_type to make sure it's cipher was selected. If it was not, it now returns none before running the logic.